fix critical bug in SymmetricJoystick alias add Heartbeat parent add doc for Heartbeat add tests for Heartbeat

+++ src/evaSlider.h
@@ -9,7 +9,7 @@ namespace eva
   /**
    * @brief Maps analog readings to 1000-2000 range
    */
-  template <class READER, int MIN_POS, int MAX_POS>
+  template <class READER, unsigned short MIN_POS, unsigned short MAX_POS>
   class Slider : public READER
   {
   public:
@@ -29,7 +29,7 @@ namespace eva
    * @tparam MIN_POS Minimum analog reading
    * @tparam MAX_POS Maximum analog reading
    */
-  template <int PIN, int PIN_MODE, int MIN_POS, int MAX_POS>
+  template <unsigned short PIN, int PIN_MODE, unsigned short MIN_POS, unsigned short MAX_POS>
   using PinSlider = Slider<AnalogPinReader<PIN, PIN_MODE>, MIN_POS, MAX_POS>;
 };
